backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# LLM APIs
openai==1.3.7
google-generativeai==0.3.2

# Docker
docker==6.1.3

# Report generation
reportlab==4.0.7
python-docx==1.1.0
jinja2==3.1.2

# Utilities
pydantic[email]==2.5.0
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
